frontend/requirements.txt
streamlit==1.28.0
requests==2.31.0
folium==0.14.0
streamlit-folium==0.15.0
python-dotenv==1.0.0
google-generativeai==0.3.1